requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pinecone
sentence-transformers
transformers
google-genai==1.25.0
google-auth==2.40.3
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8
langsmith==0.4.5